+++ Revo.Infrastructure/Events/Async/ProcessAsyncEventsJobHandler.cs
@@ -11,24 +11,26 @@ namespace Revo.Infrastructure.Events.Async
     {
         private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
 
-        private readonly IAsyncEventQueueBacklogWorker asyncEventQueueBacklogWorker;
+        private readonly IAsyncEventWorker asyncEventWorker;
         private readonly IAsyncEventPipelineConfiguration asyncEventPipelineConfiguration;
         private readonly IJobScheduler jobScheduler;
 
-        public ProcessAsyncEventsJobHandler(IAsyncEventQueueBacklogWorker asyncEventQueueBacklogWorker,
+        public ProcessAsyncEventsJobHandler(IAsyncEventWorker asyncEventWorker,
             IAsyncEventPipelineConfiguration asyncEventPipelineConfiguration,
             IJobScheduler jobScheduler)
         {
-            this.asyncEventQueueBacklogWorker = asyncEventQueueBacklogWorker;
+            this.asyncEventWorker = asyncEventWorker;
             this.asyncEventPipelineConfiguration = asyncEventPipelineConfiguration;
             this.jobScheduler = jobScheduler;
         }
 
         public async Task HandleAsync(ProcessAsyncEventsJob job, CancellationToken cancellationToken)
         {
+            Logger.Trace($"Starting to process async event queue '{job.QueueName}' with {job.AttemptsLeft} attempts left");
+
             try
             {
-                await asyncEventQueueBacklogWorker.RunQueueBacklogAsync(job.QueueName);
+                await asyncEventWorker.RunQueueBacklogAsync(job.QueueName);
             }
             catch (AsyncEventProcessingSequenceException e)
             {
@@ -37,22 +39,34 @@ namespace Revo.Infrastructure.Events.Async
             }
             catch (OptimisticConcurrencyException e)
             {
-                Logger.Debug(e, $"OptimisticConcurrencyException occurred during asynchronous queue processing");
+                Logger.Warn(e, $"Optimistic concurrency exception occurred while processing '{job.QueueName}' async event queue");
                 await ScheduleRetryAsync(job);
             }
+            catch (Exception e)
+            {
+                Logger.Error(e, $"Unhandled exception occurred while processing '{job.QueueName}' async event queue");
+                throw;
+            }
+
+            // TODO handle transient I/O errors with retries
+            Logger.Trace($"Finished processing async event queue '{job.QueueName}'");
         }
 
         private async Task ScheduleRetryAsync(ProcessAsyncEventsJob job)
         {
             if (job.AttemptsLeft > 1)
             {
-                ProcessAsyncEventsJob newJob = new ProcessAsyncEventsJob(job.QueueName, job.AttemptsLeft - 1,
-                    TimeSpan.FromTicks(job.RetryTimeout.Ticks * asyncEventPipelineConfiguration.AsyncProcessRetryTimeoutMultiplier));
+
+                TimeSpan timeout = TimeSpan.FromTicks(job.RetryTimeout.Ticks *
+                                                      asyncEventPipelineConfiguration.AsyncProcessRetryTimeoutMultiplier);
+                Logger.Debug($"Scheduling '{job.QueueName}' async event queue processing retry in {timeout.TotalSeconds} seconds");
+
+                ProcessAsyncEventsJob newJob = new ProcessAsyncEventsJob(job.QueueName, job.AttemptsLeft - 1, timeout);
                 await jobScheduler.EnqeueJobAsync(newJob, job.RetryTimeout);
             }
             else
             {
-                Logger.Error($"Not able to finish {job.QueueName} async event queue even with {asyncEventPipelineConfiguration.AsyncProcessAttemptCount} attempts, giving up");
+                Logger.Error($"Unable to finish '{job.QueueName}' async event queue even after {asyncEventPipelineConfiguration.AsyncProcessAttemptCount} attempts, giving up");
             }
         }
     }